Task_1 and Task_2 updated

+++ Task_2.c
@@ -32,6 +32,7 @@ the data in the printed output is delimited using one whitespace character.
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 
 struct node
 {
@@ -44,16 +45,15 @@ void insert_node(struct node** treePtr, int data);
 void inOrder(struct node* treePtr);
 void delete_tree(struct node** treePtr);
 
-int main() {
-	int temp = 0;
+int main(int argc, char* argv[]) 
+{
 	struct node* treePtr = NULL;
-    printf("Enter the value of the new data member: ");
-	scanf("%d", &temp);
-    while (temp > 0)
+	char* tokenPtr = NULL;
+	tokenPtr = strtok(argv[1],",");
+    while (tokenPtr != NULL)
     {
-        insert_node(&treePtr, temp);
-        printf("Enter the value of the new data member: ");
-        scanf("%d", &temp);            
+        insert_node(&treePtr,atoi(tokenPtr)); 
+		tokenPtr = strtok(NULL,",");           
     }
     printf("Initial version of binary tree:\n");
     inOrder(treePtr);
@@ -99,7 +99,11 @@ void inOrder(struct node* treePtr)
 
 void delete_tree(struct node** treePtr)
 {
-       free(*treePtr);
-	   delete_tree(&((*treePtr)->leftPtr));
-       delete_tree(&((*treePtr)->rightPtr));
-}
+    if (*treePtr != NULL) // check if node exists
+	{
+		delete_tree(&((*treePtr)->leftPtr)); // look left 
+		delete_tree(&((*treePtr)->rightPtr)); // look right 
+		free(*treePtr); // visit
+		*treePtr = NULL;
+	}
+}
